[ANDROSDK-723] Add D2ErrorCollectionRepositoryMockIntegrationShould
