+++ core/src/androidTest/java/org/hisp/dhis/android/testapp/maintenance/D2ErrorCollectionRepositoryMockIntegrationShould.java
@@ -26,22 +26,18 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.android.core.maintenance;
+package org.hisp.dhis.android.testapp.maintenance;
 
 import com.google.common.collect.Lists;
 
-import org.hisp.dhis.android.core.category.CategoryOptionComboCategoryOptionLinkTableInfo;
-import org.hisp.dhis.android.core.category.CategoryOptionTableInfo;
-import org.hisp.dhis.android.core.common.BaseIdentifiableObjectModel;
 import org.hisp.dhis.android.core.data.database.SyncedDatabaseMockIntegrationShould;
-import org.hisp.dhis.android.core.option.OptionFields;
-import org.hisp.dhis.android.core.option.OptionSetTableInfo;
-import org.hisp.dhis.android.core.option.OptionTableInfo;
+import org.hisp.dhis.android.core.maintenance.D2Error;
+import org.hisp.dhis.android.core.maintenance.D2ErrorCode;
+import org.hisp.dhis.android.core.maintenance.D2ErrorComponent;
 import org.hisp.dhis.android.core.period.PeriodType;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
@@ -51,62 +47,7 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
 
 @RunWith(AndroidJUnit4.class)
-public class MaintenanceMockIntegrationShould extends SyncedDatabaseMockIntegrationShould {
-
-    @Test
-    public void allow_access_to_foreign_key_violations() {
-        List<ForeignKeyViolation> violations = d2.maintenanceModule().foreignKeyViolations.get();
-        assertThat(violations.size(), is(2));
-
-        ForeignKeyViolation categoryOptionComboViolation = ForeignKeyViolation.builder()
-                .toTable(CategoryOptionTableInfo.TABLE_INFO.name())
-                .toColumn(BaseIdentifiableObjectModel.Columns.UID)
-                .fromTable(CategoryOptionComboCategoryOptionLinkTableInfo.TABLE_INFO.name())
-                .fromColumn(CategoryOptionComboCategoryOptionLinkTableInfo.Columns.CATEGORY_OPTION)
-                .notFoundValue("non_existent_category_option_uid")
-                .build();
-
-        ForeignKeyViolation optionViolation = ForeignKeyViolation.builder()
-                .toTable(OptionSetTableInfo.TABLE_INFO.name())
-                .toColumn(BaseIdentifiableObjectModel.Columns.UID)
-                .fromTable(OptionTableInfo.TABLE_INFO.name())
-                .fromColumn(OptionFields.OPTION_SET)
-                .notFoundValue("non_existent_option_set_uid")
-                .fromObjectUid("non_existent_option_uid")
-                .build();
-
-        List<ForeignKeyViolation> violationsToCompare = new ArrayList<>();
-        for (ForeignKeyViolation violation : violations) {
-            violationsToCompare.add(violation.toBuilder().id(null).created(null).fromObjectRow(null).build());
-        }
-
-        assertThat(violationsToCompare.contains(categoryOptionComboViolation), is(true));
-        assertThat(violationsToCompare.contains(optionViolation), is(true));
-    }
-
-    @Test
-    public void allow_access_to_foreign_key_violations_with_children() {
-        List<ForeignKeyViolation> violations = d2.maintenanceModule().foreignKeyViolations.withAllChildren().get();
-        assertThat(violations.size(), is(2));
-    }
-
-    @Test
-    public void get_no_vulnerabilities_for_high_threshold() {
-        assertThat(d2.maintenanceModule().getPerformanceHintsService(100,
-                100).areThereVulnerabilities(), is(false));
-    }
-
-    @Test
-    public void get_vulnerabilities_for_low_threshold() {
-        assertThat(d2.maintenanceModule().getPerformanceHintsService(1,
-                1).areThereVulnerabilities(), is(true));
-    }
-
-    @Test
-    public void allow_access_to_d2_errors() {
-        List<D2Error> d2Errors = d2.maintenanceModule().d2Errors.get();
-        assertThat(d2Errors.size(), is(2));
-    }
+public class D2ErrorCollectionRepositoryMockIntegrationShould extends SyncedDatabaseMockIntegrationShould {
 
     @Test
     public void filter_d2_error_by_resource_type() {
